% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_est_cov_list}
\alias{average_est_cov_list}
\title{Helper function for averaging lists of estimates
generated in the main \code{for} loop of \code{drtmle}}
\usage{
average_est_cov_list(est_cov_list)
}
\arguments{
\item{est_cov_list}{A list with named entries \code{est} and \code{cov}}
}
\description{
Helper function for averaging lists of estimates
generated in the main \code{for} loop of \code{drtmle}
}
