% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_ic_list}
\alias{average_ic_list}
\title{Helper function to average convergence results and drtmle
influence function estimates over multiple fits}
\usage{
average_ic_list(ic_list)
}
\arguments{
\item{ic_list}{List of influence function estimates}
}
\description{
Helper function to average convergence results and drtmle
influence function estimates over multiple fits
}
