% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cao.dr.R
\name{cao.dr}
\alias{cao.dr}
\title{cao.dr 
Compute the Cao 2009 estimator}
\usage{
cao.dr(R, Y, cov, family = "gaussian", nBoot = 500)
}
\arguments{
\item{R}{Missing indicator}

\item{Y}{Outcome}

\item{cov}{Covariates}

\item{family}{Character}

\item{nBoot}{Number of bootstrap resamples}
}
\description{
cao.dr 
Compute the Cao 2009 estimator
}

