% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{ci}
\alias{ci}
\title{Compute confidence intervals for drtmle and islptw@}
\usage{
ci(...)
}
\arguments{
\item{...}{Arguments to be passed to method}
}
\description{
Compute confidence intervals for drtmle and islptw@
}
