% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vermeulen2.R
\name{data.adaptive.biasreduced.DR}
\alias{data.adaptive.biasreduced.DR}
\title{data.adaptive.biasreduced.DR}
\usage{
data.adaptive.biasreduced.DR(R, Y, cov, type.initQ = c("par", "SL"),
  zeta = 0.005, fluc = c("unweighted", "weighted"), alpha = 0.05,
  psi.tilde = 0)
}
\description{
Vermeulen estimator from IJB paper
}

