% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{drtmle}
\alias{drtmle}
\title{drtmle}
\usage{
drtmle(Y, A, W, a0 = unique(A), family = "binomial", stratify = TRUE,
  libraryQ = NULL, libraryg = NULL, libraryQr = NULL, librarygr = NULL,
  glmQ = NULL, glmg = NULL, glmQr = NULL, glmgr = NULL, guard = c("Q",
  "g"), reduction = "univariate", returnModels = FALSE, maxIter = 100,
  tolEps = 1e-04, tolIC = "default", tolg = 1e-08, verbose = TRUE,
  Qsteps = 2, ...)
}
\description{
Main function to estimate TMLE with doubly-robust inference
}

