% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{estimateG}
\alias{estimateG}
\title{estimateG}
\usage{
estimateG(A, W, libraryg, glmg, family, a0, tolg, verbose = FALSE,
  returnModels = FALSE)
}
\description{
Function to estimate propensity score
}

