% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimategrn.R
\name{estimategrn}
\alias{estimategrn}
\title{estimategrn}
\usage{
estimategrn(Y, A, W, Qn, gn, librarygr, tolg, glmgr, a0, reduction)
}
\arguments{
\item{Y}{A vector of continuous or binary outcomes.}

\item{A}{A vector of binary treatment assignment (assumed to be equal to 0 or 1)}

\item{W}{A \code{data.frame} of named covariates}

\item{Qn}{A list of outcome regression estimates evaluated on observed data}

\item{gn}{A list of propensity regression estimates evaluated on observed data}

\item{librarygr}{A vector of characters or a list describing the Super Learner library to be used 
for the second reduced-dimension regression.}

\item{tolg}{A numeric indicating the minimum value for estimates of the propensity score.}

\item{glmgr}{A character describing a formula to be used in the call to \code{glm} for the second reduced-dimension regression. Ignored
if \code{librarygr!=NULL}.}

\item{a0}{A list of fixed treatment values}

\item{reduction}{A character equal to \code{"univariate"} for a univariate misspecification correction or \code{"bivariate"}
for the bivariate version.}
}
\description{
Estimates the reduced dimension regressions necessary for the additional 
fluctuations.
}

