% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{estimateQ}
\alias{estimateQ}
\title{estimateQ}
\usage{
estimateQ(Y, A, W, libraryQ, glmQ, a0, stratify, family, verbose = FALSE,
  returnModels = FALSE, ...)
}
\description{
Function to estimate initial outcome regression
}

