% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateQrn.R
\name{estimateQrn}
\alias{estimateQrn}
\title{estimateQrn}
\usage{
estimateQrn(Y, A, W, Qn, gn, glmQr, libraryQr, a0)
}
\arguments{
\item{Y}{A vector of continuous or binary outcomes.}

\item{A}{A vector of binary treatment assignment (assumed to be equal to 0 or 1)}

\item{W}{A \code{data.frame} of named covariates}

\item{Qn}{A list of outcome regression estimates evaluated on observed data}

\item{gn}{A list of propensity regression estimates evaluated on observed data}

\item{glmQr}{A character describing a formula to be used in the call to \code{glm} for the first reduced-dimension regression. Ignored
if \code{librarygr!=NULL}.}

\item{libraryQr}{A vector of characters or a list describing the Super Learner library to be used 
for the first reduced-dimension regression.}

\item{a0}{A list of fixed treatment values}
}
\description{
Estimates the reduced dimension regressions necessary for the  
fluctuations of g
}

