% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_functions.R
\name{eval_Diptw_g}
\alias{eval_Diptw_g}
\title{Evaluate extra piece of the influence function for the IPTW}
\usage{
eval_Diptw_g(A, DeltaA, DeltaY, Qrn, gn, a_0)
}
\arguments{
\item{A}{A vector of binary treatment assignment (assumed to be equal to 0 or
1)}

\item{DeltaA}{Indicator of missing treatment (assumed to be equal to 0 if
missing 1 if observed)}

\item{DeltaY}{Indicator of missing outcome (assumed to be equal to 0 if
missing 1 if observed)}

\item{Qrn}{List of estimated reduced-dimension outcome regression evaluated
at observations}

\item{gn}{List of estimated propensity scores evaluated at observations}

\item{a_0}{Vector of values to return marginal mean}
}
\description{
Evaluate extra piece of the influence function for the IPTW
}
