% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_functions.R
\name{eval_Dstar}
\alias{eval_Dstar}
\title{Evaluate usual efficient influence function}
\usage{
eval_Dstar(A, Y, DeltaY, DeltaA, Qn, gn, psi_n, a_0)
}
\arguments{
\item{A}{A vector of binary treatment assignment (assumed to be equal to 0 or
1)}

\item{Y}{A numeric of continuous or binary outcomes.}

\item{DeltaY}{Indicator of missing outcome (assumed to be equal to 0 if
missing 1 if observed)}

\item{DeltaA}{Indicator of missing treatment (assumed to be equal to 0 if
missing 1 if observed)}

\item{Qn}{List of estimated outcome regression evaluated at observations}

\item{gn}{List of estimated propensity scores evaluated at observations}

\item{psi_n}{List of estimated ATEs}

\item{a_0}{Vector of values to return marginal mean}
}
\description{
Evaluate usual efficient influence function
}
