% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_models}
\alias{extract_models}
\title{Help function to extract models from fitted object}
\usage{
extract_models(a_list)
}
\arguments{
\item{a_list}{Structured list of nuisance parameters}
}
\description{
Help function to extract models from fitted object
}
