% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{fluctuateG}
\alias{fluctuateG}
\title{fluctuateG}
\usage{
fluctuateG(Y, A, W, a0, Qn, gn, Qrn, tolg, coefTol = 1e+07)
}
\description{
Function called internally by drtmle to perform the fluctuation 
of the initial estimator of g (i.e., solves the new estimating eqn that results
from misspecification of Q)
}

