% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{fluctuateQ}
\alias{fluctuateQ}
\title{fluctuateQ}
\usage{
fluctuateQ(Y, A, W, Qn, gn, grn, a0, family, reduction, coefTol = 1e+07)
}
\description{
Function called internally by drtmle to perform simultaneous fluctuation 
of the initial estimator of Q (i.e., solves both EIF estimating eqn and 
the new estimating eqn that results from misspecification of g)
}

