% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{fluctuateQ1}
\alias{fluctuateQ1}
\title{fluctuateQ1}
\usage{
fluctuateQ1(Y, A, W, Qn, gn, a0, family)
}
\description{
Function called internally by drtmle to perform the first fluctuation 
of the initial estimator of Q (i.e., solves the original EIF estimating eqn)
}

