% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{fluctuateQ2}
\alias{fluctuateQ2}
\title{fluctuateQ2}
\usage{
fluctuateQ2(Y, A, W, Qn, gn, grn, a0, family, reduction, coefTol = 1e+07)
}
\description{
Function called internally by drtmle to perform the second fluctuation 
of the initial estimator of Q (i.e., solves the new estimating eqn that results
from misspecification of g)
}

