% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCaoEst.R
\name{getCaoEst}
\alias{getCaoEst}
\title{getCaoEst}
\usage{
getCaoEst(R, Y, cov, family)
}
\arguments{
\item{R}{Missing indicator}

\item{Y}{outcome}

\item{cov}{Covariates}

\item{family}{Character for family of outcome}
}
\description{
Code to implement Cao 2009 estimator
}

