% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vermeulen2.R
\name{m.biasreducedDR.identity}
\alias{m.biasreducedDR.identity}
\title{m.biasreducedDR.identity}
\usage{
m.biasreducedDR.identity(R, Y, cov)
}
\description{
Code to implement Vermeulen non-data-adaptive estimator
}

