% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_validRows}
\alias{make_validRows}
\title{Make list of rows in each validation fold.}
\usage{
make_validRows(cvFolds, n, ...)
}
\arguments{
\item{cvFolds}{Numeric number of cv folds}

\item{n}{Number of observations}

\item{...}{Other arguments}
}
\description{
Make list of rows in each validation fold.
}
