% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{makeDataBiv}
\alias{makeDataBiv}
\title{makeDataBiv}
\usage{
makeDataBiv(n, beta = 1)
}
\arguments{
\item{n}{Sample size}

\item{beta}{Strength of interaction}
}
\description{
Make simulation data set of size n
}

