% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_fixes.R
\name{method.CC_LS_mod}
\alias{method.CC_LS_mod}
\title{Modification of convex combination least squares method
for \code{SuperLearner} that doesn't fail with duplicated predictors.
Issue reported to \code{SuperLearner} maintainers. This function will
be deprecated when a more robust fix in the \code{SuperLearner} package
is implemented.}
\usage{
method.CC_LS_mod()
}
\description{
Modification of convex combination least squares method
for \code{SuperLearner} that doesn't fail with duplicated predictors.
Issue reported to \code{SuperLearner} maintainers. This function will
be deprecated when a more robust fix in the \code{SuperLearner} package
is implemented.
}
