% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.drtmle}
\alias{plot.drtmle}
\title{Plot reduced dimension regression fits}
\usage{
\method{plot}{drtmle}(x, nPoints = 500, a_0 = x$a_0[1], ...)
}
\arguments{
\item{x}{An object of class \code{"drtmle"}}

\item{nPoints}{Number of points to plot lines (increase for less bumpy plot,
decrease for faster evaluation).}

\item{a_0}{For what value of a_0 should the plot be made for?}

\item{...}{More arguments (not currently used) TO DO: Pass to plot? to lines?}
}
\description{
Plot reduced dimension regression fits
}
