% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.adaptive_iptw}
\alias{print.adaptive_iptw}
\title{Print the output of a \code{"adaptive_iptw"} object.}
\usage{
\method{print}{adaptive_iptw}(x, ...)
}
\arguments{
\item{x}{A \code{"adaptive_iptw"} object.}

\item{...}{Other arguments (not used)}
}
\description{
Print the output of a \code{"adaptive_iptw"} object.
}
