% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ci.adaptive_iptw}
\alias{print.ci.adaptive_iptw}
\title{Print the output of ci.adaptive_iptw}
\usage{
\method{print}{ci.adaptive_iptw}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class ci.adaptive_iptw}

\item{digits}{Number of digits to round to}

\item{...}{Other options (not currently used)}
}
\description{
Print the output of ci.adaptive_iptw
}
