% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ci.drtmle}
\alias{print.ci.drtmle}
\title{Print the output of ci.drtmle}
\usage{
\method{print}{ci.drtmle}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class ci.drtmle}

\item{digits}{Number of digits to round to}

\item{...}{Other options (not currently used)}
}
\description{
Print the output of ci.drtmle
}
