% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ci.islptw}
\alias{print.ci.islptw}
\title{Print the output of ci.islptw}
\usage{
\method{print}{ci.islptw}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class ci.islptw}

\item{digits}{Number of digits to round to}

\item{...}{Other options (not currently used)}
}
\description{
Print the output of ci.islptw
}
