% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.confint.drtmle}
\alias{print.confint.drtmle}
\title{Print the output of confint.drtmle}
\usage{
\method{print}{confint.drtmle}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class drconfint}

\item{digits}{Number of digits to round to}

\item{...}{Other options (not currently used)}
}
\description{
Print the output of confint.drtmle
}
