% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.islptw}
\alias{print.islptw}
\title{Print the output of a \code{"islptw"} object.}
\usage{
\method{print}{islptw}(x, ...)
}
\arguments{
\item{x}{A \code{"islptw"} object.}

\item{...}{Other arguments (not used)}
}
\description{
Print the output of a \code{"islptw"} object.
}
