% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.wald_test.adaptive_iptw}
\alias{print.wald_test.adaptive_iptw}
\title{Print the output of wald_test.adaptive_iptw}
\usage{
\method{print}{wald_test.adaptive_iptw}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class wald_test.adaptive_iptw}

\item{digits}{Number of digits to round to}

\item{...}{Other options (not currently used)}
}
\description{
Print the output of wald_test.adaptive_iptw
}
