% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.wald_test.islptw}
\alias{print.wald_test.islptw}
\title{Print the output of wald_test.islptw}
\usage{
\method{print}{wald_test.islptw}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class wald_test.islptw}

\item{digits}{Number of digits to round to}

\item{...}{Other options (not currently used)}
}
\description{
Print the output of wald_test.islptw
}
