% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.caret1.R
\name{SL.caret1}
\alias{SL.caret1}
\title{SL.caret1}
\usage{
SL.caret1(Y, X, newX, family, obsWeights, method = "rf", tuneLength = 3,
  trControl = caret::trainControl(method = "cv", number = ifelse(length(Y) <=
  100, 20, 10), verboseIter = FALSE), ...)
}
\arguments{
\item{Y}{Training outcomes}

\item{X}{Training predictors}

\item{newX}{Test set predictors}

\item{family}{Character indicating family argument (ignored)}

\item{obsWeights}{Weights for the observations}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}

\item{...}{Other arguments (not currently used)}
}
\description{
A modification of the SL.caret function from the SuperLearner package that
suppresses some output when method = "gbm" or "nnet". See \code{?SL.caret}
for more information.
}

