% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slWrappers.R
\name{SL.caretMod}
\alias{SL.caretMod}
\title{SL.caretMod}
\usage{
SL.caretMod(Y, X, newX, family, obsWeights, method, tuneLength = 10,
  trControl = caret::trainControl(method = "cv", number = 5, verboseIter =
  FALSE), ...)
}
\arguments{
\item{Y}{Training outcomes}

\item{X}{Training predictors}

\item{newX}{Test set predictors}

\item{family}{Character indicating family argument (ignored)}

\item{obsWeights}{Weights for the observations}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}

\item{...}{Other arguments (not currently used)}
}
\description{
A modification of the SL.caret function from the SuperLearner package that
suppresses some output when method = "gbm" or "nnet". See \code{?SL.caret}
for more information.
}
