% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slWrappers.R
\name{SL.gbm.caretMod}
\alias{SL.gbm.caretMod}
\title{SL.gbm.caretMod}
\usage{
SL.gbm.caretMod(..., method = "gbm", tuneLength = 20,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to \code{SL.caretMod}}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
Uses SL.caretMod to train a gbm with 10 choices of tuning parameters and 5-fold CV
}
