% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{SL.glmnet.caret1}
\alias{SL.glmnet.caret1}
\title{SL.svmLinear.caret1}
\usage{
SL.glmnet.caret1(..., method = "glmnet", tuneLength = 8)

SL.glmnet.caret1(..., method = "glmnet", tuneLength = 8)
}
\description{
Uses SL.caret1 to train a linear svm

Uses SL.caret1 to train a elastic net regression
}

