% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.nnet.caret1.R
\name{SL.nnet.caret1}
\alias{SL.nnet.caret1}
\title{SL.nnet.caret1}
\usage{
SL.nnet.caret1(..., method = "nnet", tuneLength = 8,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to \code{SL.caret1}}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
Uses SL.caret1 to train a neural net
}

