% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slWrappers.R
\name{SL.nnet.caretMod}
\alias{SL.nnet.caretMod}
\title{SL.nnet.caretMod}
\usage{
SL.nnet.caretMod(..., method = "nnet", tuneLength = 20,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to \code{SL.caretMod}}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
Uses SL.caretMod to train with five-fold CV to train a neural network
with twenty choices of tuning parameter.
}
