% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.npreg.R
\name{SL.npreg}
\alias{SL.npreg}
\title{SL.npreg
SuperLearner wrapper to fit kernel regression}
\usage{
SL.npreg(Y, X, newX, family, obsWeights, rangeThresh = 1e-07, ...)
}
\arguments{
\item{Y}{Training outcomes}

\item{X}{Training predictors}

\item{newX}{Test set predictors}

\item{family}{Not used by the function, but needed by \code{SuperLearner}}

\item{obsWeights}{Weights for the observations}

\item{rangeThresh}{If the the range of the outcomes is smaller than this number, the method
returns the empirical average of the outcomes.}

\item{...}{Other arguments (not currently used)}
}
\description{
SL.npreg
SuperLearner wrapper to fit kernel regression
}

