% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slWrappers.R
\name{SL.randomGLM.caretMod}
\alias{SL.randomGLM.caretMod}
\title{SL.randomGLM.caretMod
Uses SL.caretMod with five-fold CV to train a random GLM
with ten choices of tuning parameter.}
\usage{
SL.randomGLM.caretMod(..., method = "randomGLM", tuneLength = 10,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to \code{SL.caretMod}}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
SL.randomGLM.caretMod
Uses SL.caretMod with five-fold CV to train a random GLM
with ten choices of tuning parameter.
}
