% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.rf.caret1.R
\name{SL.rf.caret1}
\alias{SL.rf.caret1}
\title{SL.rf.caret1}
\usage{
SL.rf.caret1(..., method = "rf", tuneLength = 8,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to \code{SL.caret1}

Uses SL.caret1 to train random forest}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
Uses SL.caret 1 to train a gbm with 10 choices of tuning parameters and 5-fold CV
}

