% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.rf.caret2.R
\name{SL.rf.caret2}
\alias{SL.rf.caret2}
\title{SL.rf.caret2}
\usage{
SL.rf.caret2(..., method = "rf", tuneLength = 20,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to \code{SL.caret1}}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
Uses SL.caret 1 to train a random forest with 10 choices of tuning parameters and 5-fold CV
}

