% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.rpart.caret1.R
\name{SL.rpart.caret1}
\alias{SL.rpart.caret1}
\title{SL.rpart.caret1}
\usage{
SL.rpart.caret1(..., method = "rpart", tuneLength = 8,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to SL.caret1}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
Uses SL.caret1 to train regression tree.
}

