% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slWrappers.R
\name{SL.rpart.caretMod}
\alias{SL.rpart.caretMod}
\title{SL.rpart.caretMod}
\usage{
SL.rpart.caretMod(..., method = "rpart", tuneLength = 10,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to SL.caretMod}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
Uses SL.caretMod with five-fold CV to train regression tree 
with ten choices of tuning parameter.
}
