% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slWrappers.R
\name{SL.svmLinear.caretMod}
\alias{SL.svmLinear.caretMod}
\title{SL.svmLinear.caretMod
Uses SL.caretMod to train a linear svm with 15 choices of 
tuning parameter}
\usage{
SL.svmLinear.caretMod(..., method = "svmLinear", tuneLength = 15,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to \code{SL.caretMod}}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
SL.svmLinear.caretMod
Uses SL.caretMod to train a linear svm with 15 choices of 
tuning parameter
}
