% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tmp_method.CC_nloglik}
\alias{tmp_method.CC_nloglik}
\title{Temporary fix for convex combination method negative log-likelihood loss
Relative to existing implementation, we reduce the tolerance at which 
we declare predictions from a given algorithm the same as another}
\usage{
tmp_method.CC_nloglik()
}
\description{
Temporary fix for convex combination method negative log-likelihood loss
Relative to existing implementation, we reduce the tolerance at which 
we declare predictions from a given algorithm the same as another
}
