% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{wald_test}
\alias{wald_test}
\title{Wald tests for drtmle and adaptive_iptw objects}
\usage{
wald_test(...)
}
\arguments{
\item{...}{Arguments to be passed to method}
}
\description{
Wald tests for drtmle and adaptive_iptw objects
}
