% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.adaptive.biasreduced.DR.R
\name{data.adaptive.biasreduced.DR}
\alias{data.adaptive.biasreduced.DR}
\title{data.adaptive.biasreduced.DR
Vermeulen estimator from IJB paper}
\usage{
data.adaptive.biasreduced.DR(R, Y, cov, type.initQ = c("par", "SL"),
  zeta = 0.005, fluc = c("unweighted", "weighted"), alpha = 0.05,
  psi.tilde = 0)
}
\arguments{
\item{R}{Missingness indicator}

\item{Y}{Outcome}

\item{cov}{Covariates}

\item{type.initQ}{How to estimate outcome regression (\code{"par","SL","npreg"})}

\item{zeta}{Truncation level for logit transform of Qn}

\item{fluc}{How to fit fluctuation submodel}

\item{alpha}{Truncation level for propensity score gn}

\item{psi.tilde}{Null hypothesis value}
}
\description{
data.adaptive.biasreduced.DR
Vermeulen estimator from IJB paper
}

