% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtmle.R
\name{drtmle}
\alias{drtmle}
\title{drtmle}
\usage{
drtmle(Y, A, W, a0 = unique(A), family = binomial(), stratify = TRUE,
  libraryQ = NULL, libraryg = NULL, libraryQr = NULL, librarygr = NULL,
  glmQ = NULL, glmg = NULL, glmQr = NULL, glmgr = NULL, guard = c("Q",
  "g"), reduction = "univariate", returnModels = FALSE, maxIter = 10,
  tolIC = 1/length(Y), tolg = 1e-08, verbose = FALSE, Qsteps = 2, ...)
}
\arguments{
\item{Y}{A numeric of continuous or binary outcomes.}

\item{A}{A vector of binary treatment assignment (assumed to be equal to 0 or 1)}

\item{W}{A \code{data.frame} of named covariates}

\item{a0}{A vector of treatment levels at which to compute the adjusted mean outcome.}

\item{family}{A \code{character} equal to either \code{"binomial"} or \code{"gaussian"}, to be passed
to the \code{SuperLearner} or \code{glm} function.}

\item{stratify}{A \code{boolean} indicating whether to estimate the outcome regression separately
for observations with \code{A} equal to 0/1 (if \code{TRUE}) or to pool across \code{A} (if \code{FALSE}).}

\item{libraryQ}{A vector of characters or a list describing the Super Learner library to be used 
for the outcome regression. See \code{link{SuperLearner::SuperLearner}} for details.}

\item{libraryg}{A vector of characters or a list describing the Super Learner library to be used 
for the propensity score. See \code{link{SuperLearner::SuperLearner}} for details.}

\item{libraryQr}{A vector of characters or a list describing the Super Learner library to be used 
for the first reduced-dimension regression.}

\item{librarygr}{A vector of characters or a list describing the Super Learner library to be used 
for the second reduced-dimension regression.}

\item{glmQ}{A character describing a formula to be used in the call to \code{glm} for the outcome regression. Ignored
if \code{libraryQ!=NULL}.}

\item{glmg}{A character describing a formula to be used in the call to \code{glm} for the propensity score. Ignored
if \code{libraryg!=NULL}.}

\item{glmQr}{A character describing a formula to be used in the call to \code{glm} for the first reduced-dimension regression. Ignored
if \code{libraryQr!=NULL}. The formula should use the variable name \code{'gn'}.}

\item{glmgr}{A character describing a formula to be used in the call to \code{glm} for the second reduced-dimension regression. Ignored
if \code{librarygr!=NULL}. The formula should use the variable name \code{'Qn'} and \code{'gn'} if 
\code{reduction='bivariate'} and \code{'Qn'} otherwise.}

\item{guard}{A character vector indicating what pattern of misspecifications to guard against. If \code{guard} contains \code{"Q"}, 
then the TMLE guards against misspecification of the outcome regression by estimating the reduced dimension regression
specified by \code{glmQr} or \code{libraryQr}. If \code{guard} contains \code{"g"} then the TMLE 
(additionally) guards against misspecification of the propensity score by estimating the reduced dimension regression
specified by \code{glmgr} or \code{librarygr}. If \code{NULL}, the usual TMLE is computed.}

\item{reduction}{A character equal to \code{"univariate"} for a univariate misspecification correction or \code{"bivariate"}
for the bivariate version.}

\item{returnModels}{A boolean indicating whether to return model fits for the outcome regression, propensity score,
and reduced-dimension regressions.}

\item{maxIter}{A numeric that sets the maximum number of iterations the TMLE can perform in its fluctuation step.}

\item{tolIC}{A numeric that defines the stopping criteria based on the empirical mean
of the scores of the fluctuation submodels submodels. Setting to \code{"default"}}

\item{tolg}{A numeric indicating the minimum value for estimates of the propensity score.}

\item{verbose}{A boolean indicating whether to print status updates.}

\item{Qsteps}{A numeric equal to 1/2 indicating whether the fluctuation submodel for the outcome regression
should be fit using a single minimization (\code{Qsteps = 1}) or a backfitting-type minimization (\code{Qsteps=2}). 
The latter was found to be more stable in simulations.}

\item{...}{Other options (not currently used)}
}
\description{
Main function to estimate TMLE with doubly-robust inference
}
