% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateG}
\alias{estimateG}
\title{estimateG}
\usage{
estimateG(A, W, libraryg, glmg, a0, tolg, verbose = FALSE,
  returnModels = FALSE)
}
\arguments{
\item{A}{A vector of binary treatment assignment (assumed to be equal to 0 or 1)}

\item{W}{A \code{data.frame} of named covariates}

\item{libraryg}{A vector of characters or a list describing the Super Learner library to be used 
for the propensity score}

\item{glmg}{A character describing a formula to be used in the call to \code{glm} for the propensity score}

\item{a0}{A list of fixed treatment values}

\item{tolg}{A numeric indicating the minimum value for estimates of the propensity score.}

\item{verbose}{A boolean indicating whether to print status updates.}

\item{returnModels}{A boolean indicating whether to return model fits for the outcome regression, propensity score,
and reduced-dimension regressions.}
}
\description{
Function to estimate propensity score
}
