% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{estimategrn}
\alias{estimategrn}
\title{estimategrn}
\usage{
estimategrn(Y, A, W, Qn, gn, librarygr, tolg, glmgr, a0, reduction)
}
\description{
Estimates the reduced dimension regressions necessary for the additional 
fluctuations.
}

