% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{estimateQrn}
\alias{estimateQrn}
\title{estimateQrn}
\usage{
estimateQrn(Y, A, W, Qn, gn, glmQr, libraryQr, a0)
}
\description{
Estimates the reduced dimension regressions necessary for the  
fluctuations of g
}

