% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluctuate.R
\name{fluctuateQ}
\alias{fluctuateQ}
\title{fluctuateQ}
\usage{
fluctuateQ(
  Y,
  A,
  W,
  DeltaY,
  DeltaA,
  Qn,
  gn,
  grn,
  a_0,
  reduction,
  coefTol = 1000
)
}
\arguments{
\item{Y}{The outcome}

\item{A}{The treatment}

\item{W}{The covariates}

\item{DeltaY}{Indicator of missing outcome (assumed to be equal to 0 if
missing 1 if observed)}

\item{DeltaA}{Indicator of missing treatment (assumed to be equal to 0 if
missing 1 if observed)}

\item{Qn}{A list of outcome regression estimates evaluated on observed data}

\item{gn}{A list of propensity regression estimates evaluated on observed
data}

\item{grn}{A list of reduced-dimension regression estimates evaluated on
observed data}

\item{a_0}{A list of fixed treatment values}

\item{reduction}{A character indicating what reduced dimension regression was
used.}

\item{coefTol}{A tolerance level on the magnitude of the coefficient that
flags the result as potentially the result of numeric instability.}
}
\description{
Function called internally by drtmle to perform simultaneous fluctuation
of the initial estimator of Q (i.e., solves both EIF estimating eqn and
the new estimating eqn that results from misspecification of g)
}
