% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluctuate.R
\name{fluctuateQ1}
\alias{fluctuateQ1}
\title{fluctuateQ1}
\usage{
fluctuateQ1(Y, A, W, DeltaA, DeltaY, Qn, gn, a_0)
}
\arguments{
\item{Y}{The outcome}

\item{A}{The treatment}

\item{W}{The covariates}

\item{DeltaA}{Indicator of missing treatment (assumed to be equal to 0 if missing 1 if observed)}

\item{DeltaY}{Indicator of missing outcome (assumed to be equal to 0 if missing 1 if observed)}

\item{Qn}{A list of outcome regression estimates evaluated on observed data}

\item{gn}{A list of propensity regression estimates evaluated on observed data}

\item{a_0}{A list of fixed treatment values}
}
\description{
Function called internally by drtmle to perform the first fluctuation 
of the initial estimator of Q (i.e., solves the original EIF estimating eqn)
}
