% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluctuateQ2.R
\name{fluctuateQ2}
\alias{fluctuateQ2}
\title{fluctuateQ2}
\usage{
fluctuateQ2(Y, A, W, Qn, gn, grn, a0, reduction, coefTol = 1e+05)
}
\arguments{
\item{Y}{The outcome}

\item{A}{The treatment}

\item{W}{The covariates}

\item{Qn}{A list of outcome regression estimates evaluated on observed data}

\item{gn}{A list of propensity regression estimates evaluated on observed data}

\item{grn}{A list of reduced-dimension regression estimates evaluated on observed data}

\item{a0}{A list of fixed treatment values}

\item{reduction}{A character indicating what reduced dimension regression was used.}

\item{coefTol}{A tolerance level on the magnitude of the coefficient that flags the
result as potentially the result of numeric instability.}
}
\description{
Function called internally by drtmle to perform the second fluctuation 
of the initial estimator of Q (i.e., solves the new estimating eqn that results
from misspecification of g)
}

