% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.biasreducedDR.identity.R
\name{m.biasreducedDR.identity}
\alias{m.biasreducedDR.identity}
\title{m.biasreducedDR.identity}
\usage{
m.biasreducedDR.identity(R, Y, cov)
}
\arguments{
\item{R}{Missingness indicator}

\item{Y}{Outcome}

\item{cov}{Covariates}
}
\description{
Code to implement Vermeulen non-data-adaptive estimator
}

