% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slWrappers.R
\name{predict.SL.npreg}
\alias{predict.SL.npreg}
\title{predict.SL.npreg}
\usage{
\method{predict}{SL.npreg}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{"SL.npreg"}}

\item{newdata}{The new data used to obtain predictions}

\item{...}{Other arguments passed to predict}
}
\description{
Method for predicting SL.npreg objects.
}
