% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.drtmle}
\alias{print.drtmle}
\title{Print the output of a \code{"drtmle"} object.}
\usage{
\method{print}{drtmle}(x, ...)
}
\arguments{
\item{x}{A \code{"drtmle"} object}

\item{...}{Other arguments (not used)}
}
\description{
Print the output of a \code{"drtmle"} object.
}
