% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{SL.caret1}
\alias{SL.caret1}
\title{SL.caret1}
\usage{
SL.caret1(Y, X, newX, family, obsWeights, method = "rf", tuneLength = 3,
  trControl = caret::trainControl(method = "cv", number = ifelse(length(Y) <=
  100, 20, 10), verboseIter = FALSE), metric, ...)
}
\description{
A modification of the SL.caret function from the SuperLearner package that
suppresses some output when method = "gbm".
}

