% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{SL.gbm.caret2}
\alias{SL.gbm.caret2}
\title{SL.gbm.caret2
Uses SL.caret 1 to train a gbm with 10 choices of tuning parameters and 5-fold CV}
\usage{
SL.gbm.caret2(..., method = "gbm", tuneLength = 20,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\description{
SL.gbm.caret2
Uses SL.caret 1 to train a gbm with 10 choices of tuning parameters and 5-fold CV
}

