% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_npreg.R
\name{SL.npreg}
\alias{SL.npreg}
\title{Super learner wrapper for kernel regression}
\usage{
SL.npreg(Y, X, newX, family, obsWeights, rangeThresh = 1e-07, ...)
}
\arguments{
\item{Y}{Training outcomes}

\item{X}{Training predictors}

\item{newX}{Test set predictors}

\item{family}{Not used by the function, but needed for compatibility with \code{SuperLearner}}

\item{obsWeights}{Weights for the observations}

\item{rangeThresh}{For stability: if the the range of the outcomes is smaller than this number, the method
returns the empirical average of the outcomes.}

\item{...}{Other arguments (not currently used)}
}
\description{
Kernel regression based on the \href{https://cran.r-project.org/web/packages/np/}{np}
package. Uses leave-one-out cross-validation to fit a kernel regression. 
See \code{?npreg} for more details.
}
