% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{SL.randomGLM.caret1}
\alias{SL.randomGLM.caret1}
\title{SL.randomGLM.caret1}
\usage{
SL.randomGLM.caret1(..., method = "randomGLM", tuneLength = 8)
}
\description{
Uses SL.caret1 to train a random GLM
}

