% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{SL.rf.caret1}
\alias{SL.rf.caret1}
\title{SL.rf.caret1}
\usage{
SL.rf.caret1(..., method = "rf", tuneLength = 8,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\description{
Uses SL.caret1 to train random forest
}

