% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{SL.rf.caret2}
\alias{SL.rf.caret2}
\title{SL.rf.caret2
Uses SL.caret 1 to train a random forest with 10 choices of tuning parameters and 5-fold CV}
\usage{
SL.rf.caret2(..., method = "rf", tuneLength = 20,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\description{
SL.rf.caret2
Uses SL.caret 1 to train a random forest with 10 choices of tuning parameters and 5-fold CV
}

