% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{SL.rpart.caret1}
\alias{SL.rpart.caret1}
\title{SL.rpart.caret1}
\usage{
SL.rpart.caret1(..., method = "rpart", tuneLength = 8,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\description{
Uses SL.caret1 to train rpart
}

