% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.svmLinear.caret1.R
\name{SL.svmLinear.caret1}
\alias{SL.svmLinear.caret1}
\title{SL.svmLinear.caret1 
Uses SL.caret1 to train a linear svm}
\usage{
SL.svmLinear.caret1(..., method = "svmLinear", tuneLength = 8,
  trControl = caret::trainControl(method = "cv", number = 5))
}
\arguments{
\item{...}{Other arguments passed to \code{SL.caret1}}

\item{method}{Character describing what algorithm to train}

\item{tuneLength}{The number of tuning parameter combinations}

\item{trControl}{Passed to \code{caret::train}}
}
\description{
SL.svmLinear.caret1 
Uses SL.caret1 to train a linear svm
}

